+++ ProjectM-Transaction-SMS/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0](https://github.com/saurabhgupta050890/transaction-sms-parser/compare/v2.3.5...v3.0.0) (2023-11-28)
+
+
+### ⚠ BREAKING CHANGES
+
+* added reference no and merchant information
+
+### Features
+
+* added reference no and merchant information ([6a7fe11](https://github.com/saurabhgupta050890/transaction-sms-parser/commit/6a7fe11fba78f6ad314679878a3044786868fd21))
+
+### [2.3.5](https://github.com/saurabhgupta050890/transaction-sms-parser/compare/v2.3.4...v2.3.5) (2023-11-28)
+
+
+### Bug Fixes
+
+* eslint plugin fixes ([c313204](https://github.com/saurabhgupta050890/transaction-sms-parser/commit/c313204f7ba2888ccfef4d6c23abf32cc31c6e5e))
+
+### [2.3.4](https://github.com/saurabhgupta050890/transaction-sms-parser/compare/v3.1.0...v2.3.4) (2023-11-28)
+
 ### [2.3.3](https://github.com/saurabhgupta050890/transaction-sms-parser/compare/v2.3.2...v2.3.3) (2023-03-12)